tests/wpt/tests/css/css-text/parsing/text-justify-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().textJustify</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-justify">
<meta name="assert" content="text-justify computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-justify", "auto");
test_computed_value("text-justify", "none");
test_computed_value("text-justify", "inter-word");
test_computed_value("text-justify", "inter-character");
</script>
</body>
</html>
